requirements.txt
torch
torchvision
pillow
streamlit
numpy


